# Backend/swagger.yaml
openapi: 3.0.3
info:
  title: Book Marketplace API
  description: API for a book marketplace application allowing users to buy, sell, and review physical, digital, and audiobook products.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://bookapi-c2zu.onrender.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the user
        name:
          type: string
          description: User's full name
        email:
          type: string
          format: email
          description: User's email address
        role:
          type: string
          enum: [seller, buyer, admin]
          description: User's role in the platform
        location:
          type: object
          properties:
            address:
              type: string
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                lng:
                  type: number
        profileImageUrl:
          type: string
          format: uri
          description: URL to the user's profile image
        isEmailConfirmed:
          type: boolean
          description: Whether the user's email is confirmed
      required:
        - name
        - email
        - role
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
      required:
        - success
        - error
...
paths:
...
  /api/user:
    post:
      summary: Create a new user
      tags: [User Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Jane Doe
                email:
                  type: string
                  format: email
                  example: jane@example.com
                password:
                  type: string
                  example: Password123
                role:
                  type: string
                  enum: [seller, buyer]
                  example: buyer
              required:
                - name
                - email
                - password
                - role
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get all users
      tags: [User Management]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...